+++ CHANGELOG.md
@@ -4,8 +4,10 @@
 
 ### Bug fixes
 
+- [#67][GL-67] - Players are set to survival mode and are able to break out of the lobby before being teleported
 - [#68][GL-68] - Spectators will still join the game if the Double Dates scenario is enabled
 
+[GL-67]: https://gitlab.com/uhccore/uhccore/-/issues/67
 [GL-68]: https://gitlab.com/uhccore/uhccore/-/issues/68
 
 ## [1.20.4] - 2023-03-24